+++ test/dloader/page/DiscographyTest.java
@@ -18,10 +18,10 @@ import dloader.pagejob.ProgressReporter;
 public class DiscographyTest {
 	class DummyProgressReporter implements ProgressReporter {
 		String repS;
-		int repI;
+		long repI;
 		
 		@Override
-		public void report(String type, int report) {
+		public void report(String type, long report) {
 			repS = type;
 			repI = report;
 		}
